lesson_29_separate_project/requirements.txt
pytest
selenium
requests
assertpy
pytest-html
psycopg2-binary
allure-pytest
